Lazy-load block-library private APIs via @wordpress/lazy-editor
Move the runtime window.wp.blockLibrary access into a reusable
useBlockLibraryPrivateApis hook in @wordpress/lazy-editor, and
re-export fetchLinkSuggestions from the same package, so route
code imports everything through the lazy-editor layer.
