+++ packages/lazy-editor/src/index.tsx
@@ -1,7 +1,13 @@
 /**
+ * WordPress dependencies
+ */
+export { __experimentalFetchLinkSuggestions as fetchLinkSuggestions } from '@wordpress/core-data';
+
+/**
  * Internal dependencies
  */
 export { Editor } from './components/editor';
 export { Preview } from './components/preview';
+export { useBlockLibraryPrivateApis } from './hooks/use-block-library-private-apis';
 export { useEditorAssets, loadEditorAssets } from './hooks/use-editor-assets';
 export { useEditorSettings } from './hooks/use-editor-settings';
