+++ routes/navigation-edit/editor/index.tsx
@@ -7,11 +7,10 @@ import { BlockEditorProvider } from '@wordpress/block-editor';
 // @ts-expect-error - No type declarations available for @wordpress/blocks
 import { createBlock } from '@wordpress/blocks';
 import { Spinner } from '@wordpress/components';
-import { useEditorAssets } from '@wordpress/lazy-editor';
 import {
-	// @ts-expect-error - No type declarations available for this export
-	__experimentalFetchLinkSuggestions as fetchLinkSuggestions,
-} from '@wordpress/core-data';
+	fetchLinkSuggestions,
+	useEditorAssets,
+} from '@wordpress/lazy-editor';
 
 /**
  * Internal dependencies
